chore: Moved ScrollableSection.tsx to the components directory
